Seperate runner.py and configuration
